chore(api): move services-utils.js to evaluation scope
